add if

+++ src/main/java/de/trawizardsOfJava/web/AppController.java
@@ -144,36 +144,46 @@ public class AppController {
 		return "backToTheFuture";
 	}
 
-	@GetMapping("/artikel/{id}/anfrage")
-	public String neueAnfrage(Model model, @PathVariable Long id, Principal principal) {
-		model.addAttribute("id", id);
-		Artikel artikel =  artikelRepository.findById(id).get();
-		ArrayList<Ausleihe> ausleihen = ausleiheRepository.findByArtikel(artikel);
-		ArrayList<Verfuegbarkeit> verbuebarkeiten = new ArrayList<>();
-		for (Ausleihe ausleihe: ausleihen){
-			verbuebarkeiten.add(ausleihe.getVerfuegbarkeit());
+	@GetMapping("/account/{benutzername}/artikel/{id}/anfrage")
+	public String neueAnfrage(Model model, @PathVariable String benutzername, @PathVariable Long id, Principal principal) {
+		if (benutzername.equals(principal.getName())) {
+			model.addAttribute("id", id);
+			Artikel artikel = artikelRepository.findById(id).get();
+			ArrayList<Ausleihe> ausleihen = ausleiheRepository.findByArtikel(artikel);
+			ArrayList<Verfuegbarkeit> verbuebarkeiten = new ArrayList<>();
+			for (Ausleihe ausleihe : ausleihen) {
+				verbuebarkeiten.add(ausleihe.getVerfuegbarkeit());
+			}
+			model.addAttribute("daten", verbuebarkeiten);
+			model.addAttribute("name", principal.getName());
+			return "ausleihe";
 		}
-		model.addAttribute("daten", verbuebarkeiten);
-		model.addAttribute("name", principal.getName());
-		return "ausleihe";
+		else {
+			return "permissionDenied";
+		}
 	}
 
-	@PostMapping("/artikel/{id}/anfrage")
-	public String speichereAnfrage(@PathVariable Long id, @RequestParam String daterange, Principal principal) {
-		Artikel artikel = artikelRepository.findById(id).get();
-		Verfuegbarkeit verfuegbarkeit = new Verfuegbarkeit();
-		verfuegbarkeit.toVerfuegbarkeit(daterange);
-		Ausleihe ausleihe = new Ausleihe();
-		ausleihe.setVerfuegbarkeit(verfuegbarkeit);
-		ausleihe.setArtikel(artikel);
-		ausleihe.setAusleihender(principal.getName());
-		ausleiheRepository.save(ausleihe);
-		Message message = new Message();
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger(artikel.getVerleiherBenutzername());
-		message.setNachricht("Anfrage für " + artikel.getArtikelName());
-		messageRepository.save(message);
-		return "backToTheFuture";
+	@PostMapping("/account/{benutzername}/artikel/{id}/anfrage")
+	public String speichereAnfrage(@PathVariable Long id, @PathVariable String benutzername,@RequestParam String daterange, Principal principal) {
+		if (benutzername.equals(principal.getName())) {
+			Artikel artikel = artikelRepository.findById(id).get();
+			Verfuegbarkeit verfuegbarkeit = new Verfuegbarkeit();
+			verfuegbarkeit.toVerfuegbarkeit(daterange);
+			Ausleihe ausleihe = new Ausleihe();
+			ausleihe.setVerfuegbarkeit(verfuegbarkeit);
+			ausleihe.setArtikel(artikel);
+			ausleihe.setAusleihender(principal.getName());
+			ausleiheRepository.save(ausleihe);
+			Message message = new Message();
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger(artikel.getVerleiherBenutzername());
+			message.setNachricht("Anfrage für " + artikel.getArtikelName());
+			messageRepository.save(message);
+			return "backToTheFuture";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
     @GetMapping("/account/{benutzername}/ausleihenuebersicht")
@@ -189,76 +199,106 @@ public class AppController {
 		}
     }
 
-    @GetMapping("/annahme/{id}")
-    public String ausleihebestaetigt(@PathVariable Long id, Model model, Principal principal){
-        Ausleihe ausleihe = ausleiheRepository.findById(id).get();
-        ausleihe.setAccepted(true);
-        ausleiheRepository.save(ausleihe);
-		Message message = new Message();
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger(ausleihe.getAusleihender());
-		message.setNachricht("Anfrage für " + ausleihe.getArtikel().getArtikelName() + " angenommen");
-		messageRepository.save(message);
-        model.addAttribute("ausleihen", ausleiheRepository.findByVerleiherName(principal.getName()));
-		model.addAttribute("name", principal.getName());
-        return "ausleihenuebersicht";
+    @GetMapping("/account/{benutzername}/annahme/{id}")
+    public String ausleihebestaetigt(@PathVariable Long id, @PathVariable String benutzername, Model model, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			Ausleihe ausleihe = ausleiheRepository.findById(id).get();
+			ausleihe.setAccepted(true);
+			ausleiheRepository.save(ausleihe);
+			Message message = new Message();
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger(ausleihe.getAusleihender());
+			message.setNachricht("Anfrage für " + ausleihe.getArtikel().getArtikelName() + " angenommen");
+			messageRepository.save(message);
+			model.addAttribute("ausleihen", ausleiheRepository.findByVerleiherName(principal.getName()));
+			model.addAttribute("name", principal.getName());
+			return "ausleihenuebersicht";
+		}
+        else {
+			return "permissionDenied";
+		}
     }
 
-    @GetMapping("/remove/{id}")
-    public String ausleiheabgelehnt(@PathVariable Long id, Model model, Principal principal){
-		Ausleihe ausleihe = ausleiheRepository.findById(id).get();
-        ausleiheRepository.delete(ausleihe);
-		Message message = new Message();
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger(ausleihe.getAusleihender());
-		message.setNachricht("Anfrage für " + ausleihe.getArtikel().getArtikelName() + " abgelehnt");
-		messageRepository.save(message);
-		model.addAttribute("ausleihen", ausleiheRepository.findByVerleiherName(principal.getName()));
-		model.addAttribute("name", principal.getName());
-        return "ausleihenuebersicht";
+    @GetMapping("/account/{benutzername}/remove/{id}")
+    public String ausleiheabgelehnt(@PathVariable Long id, Model model, @PathVariable String benutzername, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			Ausleihe ausleihe = ausleiheRepository.findById(id).get();
+			ausleiheRepository.delete(ausleihe);
+			Message message = new Message();
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger(ausleihe.getAusleihender());
+			message.setNachricht("Anfrage für " + ausleihe.getArtikel().getArtikelName() + " abgelehnt");
+			messageRepository.save(message);
+			model.addAttribute("ausleihen", ausleiheRepository.findByVerleiherName(principal.getName()));
+			model.addAttribute("name", principal.getName());
+			return "ausleihenuebersicht";
+		}
+		else {
+			return "permissionDenied";
+		}
     }
 
 	@GetMapping("/account/{benutzername}/ausgelieheneuebersicht")
-	public String leihenuebersicht(Model model, Principal principal){
-		ArrayList<Ausleihe> ausleihen = ausleiheRepository.findByAusleihender(principal.getName());
-		model.addAttribute("name", principal.getName());
-		model.addAttribute("ausleihen",ausleihen);
-		return "ausgelieheneuebersicht";
+	public String leihenuebersicht(Model model, @PathVariable String benutzername, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			ArrayList<Ausleihe> ausleihen = ausleiheRepository.findByAusleihender(principal.getName());
+			model.addAttribute("name", principal.getName());
+			model.addAttribute("ausleihen", ausleihen);
+			return "ausgelieheneuebersicht";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
 	@GetMapping("/rueckgabe/{id}")
-	public String zurueckgegeben(@PathVariable Long id, Model model, Principal principal){
-		Ausleihe ausleihe = ausleiheRepository.findById(id).get();
-		rueckgabeRepository.save(ausleihe.convertToRueckgabe());
-		Message message = new Message();
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger(ausleihe.getVerleiherName());
-		message.setNachricht(ausleihe.getArtikel().getArtikelName() + " zurückgegeben");
-		messageRepository.save(message);
-		ausleiheRepository.delete(ausleiheRepository.findById(id).get());
-		model.addAttribute("ausleihen", ausleiheRepository.findByAusleihender(principal.getName()));
-		return "ausgelieheneuebersicht";
+	public String zurueckgegeben(@PathVariable Long id, @PathVariable String benutzername, Model model, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			Ausleihe ausleihe = ausleiheRepository.findById(id).get();
+			rueckgabeRepository.save(ausleihe.convertToRueckgabe());
+			Message message = new Message();
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger(ausleihe.getVerleiherName());
+			message.setNachricht(ausleihe.getArtikel().getArtikelName() + " zurückgegeben");
+			messageRepository.save(message);
+			ausleiheRepository.delete(ausleiheRepository.findById(id).get());
+			model.addAttribute("ausleihen", ausleiheRepository.findByAusleihender(principal.getName()));
+			return "ausgelieheneuebersicht";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
 	@GetMapping("/account/{benutzername}/zurueckgegebeneartikel")
-	public String rueckgabenuebersicht(Model model, Principal principal){
-		ArrayList<Rueckgabe> ausleihen = rueckgabeRepository.findByVerleiherName(principal.getName());
-		model.addAttribute("ausleihen",ausleihen);
-		model.addAttribute("name", principal.getName());
-		return "zurueckgegebeneartikel";
+	public String rueckgabenuebersicht(Model model, @PathVariable String benutzername, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			ArrayList<Rueckgabe> ausleihen = rueckgabeRepository.findByVerleiherName(principal.getName());
+			model.addAttribute("ausleihen", ausleihen);
+			model.addAttribute("name", principal.getName());
+			return "zurueckgegebeneartikel";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
-	@GetMapping("/rueckgabe/akzeptiert/{id}")
-	public String rueckgabeakzeptiert(@PathVariable Long id, Model model, Principal principal){
-		Rueckgabe rueckgabe= rueckgabeRepository.findById(id).get();
-		Message message = new Message();
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger(rueckgabe.getVerleiherName());
-		message.setNachricht("Rückgabe von " + rueckgabe.getArtikel().getArtikelName() + " akzeptiert");
-		messageRepository.save(message);
-		rueckgabeRepository.delete(rueckgabe);
-		model.addAttribute("ausleihen", rueckgabeRepository.findByVerleiherName(principal.getName()));
-		return "zurueckgegebeneartikel";
+	@GetMapping("/account/{benutzername}/rueckgabe/akzeptiert/{id}")
+	public String rueckgabeakzeptiert(@PathVariable Long id, @PathVariable String benutzername, Model model, Principal principal){
+		if (benutzername.equals(principal.getName())) {
+			Rueckgabe rueckgabe = rueckgabeRepository.findById(id).get();
+			Message message = new Message();
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger(rueckgabe.getVerleiherName());
+			message.setNachricht("Rückgabe von " + rueckgabe.getArtikel().getArtikelName() + " akzeptiert");
+			messageRepository.save(message);
+			rueckgabeRepository.delete(rueckgabe);
+			model.addAttribute("ausleihen", rueckgabeRepository.findByVerleiherName(principal.getName()));
+			return "zurueckgegebeneartikel";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
 	@GetMapping("/account/{benutzername}/nachrichten")
@@ -273,21 +313,45 @@ public class AppController {
 		}
 	}
 
-	@PostMapping("/konflikt/send")
-	public String konfliktAbsenden(Konflikt konflikt, Principal principal){
-		konfliktRepository.save(konflikt);
-		Message message = new Message();
-		//message.setNachricht(konflikt.getBeschreibung());
-		message.setAbsender(principal.getName());
-		message.setEmpfaenger("");
-		return "benutzeransicht";
+	@GetMapping("/account/{benutzername}/konflikt/send/{id}")
+	public String konfliktAnnehmen(Model model, @PathVariable String benutzername, Principal principal, @PathVariable Long id) {
+		if (principal.getName().equals(benutzername)) {
+			Konflikt konflikt = new Konflikt();
+			konflikt.setRueckgabe(rueckgabeRepository.findById(id).get());
+			model.addAttribute("konflikt", konflikt);
+			model.addAttribute("name", principal.getName());
+			return "konfliktErstellung";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
-	@GetMapping("/nachricht/delete/{id}")
-	private String messageDelete(@PathVariable Long id, Model model, Principal principal){
-		messageRepository.delete(messageRepository.findById(id).get());
-		model.addAttribute("messages",messageRepository.findByEmpfaenger(principal.getName()));
-		return "nachrichtenUebersicht";
+	@PostMapping("/account/{benutzername}/konflikt/send")
+	public String konfliktAbsenden(Konflikt konflikt, @PathVariable String benutzername, Principal principal){
+		if (principal.getName().equals(benutzername)) {
+			konfliktRepository.save(konflikt);
+			Message message = new Message();
+			//message.setNachricht(konflikt.getBeschreibung());
+			message.setAbsender(principal.getName());
+			message.setEmpfaenger("");
+			return "benutzeransicht";
+		}
+		else {
+			return "permissionDenied";
+		}
+	}
+
+	@GetMapping("/account/{benutzername}/nachricht/delete/{id}")
+	private String messageDelete(@PathVariable Long id, @PathVariable String benutzername, Model model, Principal principal){
+		if (principal.getName().equals(benutzername)) {
+			messageRepository.delete(messageRepository.findById(id).get());
+			model.addAttribute("messages", messageRepository.findByEmpfaenger(principal.getName()));
+			return "nachrichtenUebersicht";
+		}
+		else {
+			return "permissionDenied";
+		}
 	}
 
 }
